% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_help_files.R
\name{create_help_files}
\alias{create_help_files}
\title{Create Directory of Help Files}
\usage{
create_help_files(files, help_dir = "helpfiles")
}
\arguments{
\item{files}{A character vector of names to use in creating help files.}

\item{help_dir}{A character string of the directory to use for help files.}
}
\description{
Helper function to be run interactively - it will initialise a directory of
help files (.md files) in help_dir, using the vector of files. Pass it a vector
of filenames which you which to use with \code{type = "markdown"} in your calls to
\code{\link{helper}}.
}
\examples{
create_help_files(files = c("Clusters", "Columns", "PlotHelp"), 
                  help_dir = "helpfiles")

}
