\name{epi}
\alias{epi}
\docType{data}
\title{2012 Environmental Performance Index Data}
\description{2012 Environmental Performance Index Data}
\usage{data(epi)}
\format{
  A data frame with 1452 observations on the following 45 variables.
  \describe{
    \item{\code{ISO3V10}}{a character vector}
    \item{\code{Country}}{a character vector}
    \item{\code{EPI_regions}}{a character vector}
    \item{\code{Landlock}}{a numeric vector}
    \item{\code{NoFor}}{a numeric vector}
    \item{\code{NoEle}}{a numeric vector}
    \item{\code{GDPCAP}}{a numeric vector}
    \item{\code{GDPgroup}}{a character vector}
    \item{\code{Year}}{a numeric vector}
    \item{\code{ACSATINV}}{a numeric vector}
    \item{\code{AGSUB}}{a numeric vector}
    \item{\code{AZE}}{a numeric vector}
    \item{\code{CHMORT}}{a numeric vector}
    \item{\code{CO2CAP}}{a numeric vector}
    \item{\code{CO2GDP}}{a numeric vector}
    \item{\code{CO2KWH}}{a numeric vector}
    \item{\code{EH}}{a numeric vector}
    \item{\code{EH_Air}}{a numeric vector}
    \item{\code{EH_EH}}{a numeric vector}
    \item{\code{EH_Water}}{a numeric vector}
    \item{\code{EPI}}{a numeric vector}
    \item{\code{EV}}{a numeric vector}
    \item{\code{EV_Ag}}{a numeric vector}
    \item{\code{EV_Air}}{a numeric vector}
    \item{\code{EV_BH}}{a numeric vector}
    \item{\code{EV_CC}}{a numeric vector}
    \item{\code{EV_Fish}}{a numeric vector}
    \item{\code{EV_Forest}}{a numeric vector}
    \item{\code{EV_Water}}{a numeric vector}
    \item{\code{FORCOVINV}}{a numeric vector}
    \item{\code{FORGROINV}}{a numeric vector}
    \item{\code{FORLOSS}}{a numeric vector}
    \item{\code{FSOC}}{a numeric vector}
    \item{\code{INDOOR}}{a numeric vector}
    \item{\code{MPAEEZ}}{a numeric vector}
    \item{\code{PACOV}}{a numeric vector}
    \item{\code{PM25}}{a numeric vector}
    \item{\code{POP}}{a numeric vector}
    \item{\code{POPs}}{a numeric vector}
    \item{\code{RENEW}}{a numeric vector}
    \item{\code{SO2CAP}}{a numeric vector}
    \item{\code{SO2GDP}}{a numeric vector}
    \item{\code{TCEEZ}}{a numeric vector}
    \item{\code{WATSUPINV}}{a numeric vector}
    \item{\code{WATUSEINV}}{a numeric vector}
  }
}
\details{
Please see \url{http://epi.yale.edu}.
}
\source{
Please see \url{http://epi.yale.edu}.
}
\references{
Please see publications available at \url{http://epi.yale.edu}.
}
\examples{
  # Examples here
  data(epi)
}
\keyword{datasets}
