% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper.R
\name{help_message}
\alias{help_message}
\title{Function for adding an onclick event to each helper, using shinyjs}
\usage{
help_message(id, help_dir)
}
\arguments{
\item{id}{The id of the shiny.tag this help applies to}

\item{help_dir}{The directory in which to look for help files}
}
\description{
Function for adding an onclick event to each helper, using shinyjs
}
