% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper.R
\name{observe_helpers}
\alias{observe_helpers}
\title{Function to go in server.R to use the help messages}
\usage{
observe_helpers(input, output, help_dir = "helpfiles", sizes = "m")
}
\arguments{
\item{input}{The input object in your shiny app}

\item{output}{The output object in your shiny app}

\item{help_dir}{A character string of the directory containing your helpfiles}

\item{sizes}{An optional named list of sizes for the input/output ids}
}
\description{
Function to go in server.R to use the help messages
}
\examples{
\dontrun{observe_helpers(input, output, "helpfiles")}
}
