\name{ShinyHelper-package}
\alias{ShinyHelper-package}
\alias{ShinyHelper}
\docType{package}
\title{
Making shiny application creation easier
}
\description{
This package provides a very small number of functions to
create basic shiny data exploration apps without having to directly
author the ui.R or server.R files.  An example is provided using 2012
Environmental Performance Index (EPI) from the Yale Center for
Environmental Law and Policy and Columbia's Center for International
Earth Science Information Network.
}
\details{
\tabular{ll}{
Package: \tab ShinyHelper\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-07-26\cr
License: \tab GPL-3\cr
}
Package overview goes here.
}
\author{
John W. Emerson

Maintainer: Jay Emerson <john.emerson@yale.edu>
}
\references{
REFERENCES
}
\keyword{ package }
\seealso{
\code{\link[shiny:shiny-package]{shiny}}
}
\examples{
# Examples go here
}
