% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper-package.r
\docType{package}
\name{shinyhelper}
\alias{shinyhelper}
\title{shinyhelper.}
\description{
Easily add markdown helpfiles to your shiny apps, by adding a single function
around a shiny app element. This adds a small icon (a question mark, by default)
which presents content in a modal dialog box when clicked. The content can be passed 
inline or given in a markdown file.
}
\details{
There is also a convenient function to initialise a directory of markdown files
given a vector of filenames.

A demo is available at \url{https://cwthom94.shinyapps.io/shinyhelper-demo/}

For further reference on how to use the package, please refer to
\url{https://github.com/cwthom/shinyhelper}
}
