% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper_demo.R
\name{shinyhelper_demo}
\alias{shinyhelper_demo}
\title{Launch the Example App}
\usage{
shinyhelper_demo()
}
\description{
An example app demonstrating the use of the \code{shinyhelper} package.
}
\examples{
\dontrun{
if (interactive()) {
  shinyhelper_demo()
}
}

}
