% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper-demo.R
\name{shinyhelperDemo}
\alias{shinyhelperDemo}
\title{Launch the Example App}
\usage{
shinyhelperDemo()
}
\description{
An example app demonstrating the use of the \code{shinyhelper} package.
}
\examples{
\dontrun{
if (interactive()) {
  shinyhelperDemo()
}
}

}
