\name{shinySkeleton}
\alias{shinySkeleton}
\title{
  Create a skeleton for a Shiny application
}
\description{
  Create a skeleton for a Shiny application
}
\usage{
shinySkeleton(name = "myshinyapp")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{name}{the application name (will be a folder with ui.R and server.R)}
}
\details{
  Not much to say
}
\value{
  Nothing is returned.
}

\seealso{
  Like \code{\link{package.skeleton}} I suppose
}
\examples{
  # None at the moment
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ package }
