\name{showUI}
\alias{showUI}
\alias{showServer}
\alias{showData}
\alias{flushShinyApp}
\title{
Convenience functions for the Shiny application creation process
}
\description{
Explore members of the Shiny application under construction, or
flush the application before starting from scratch.
}
\usage{
showUI()
showServer()
showData()
flushShinyApp()
}
\details{
These functions basically aid the interaction with .GlobalEnv$shinier.
}
\value{
Functionality is obvious
}
\references{
None
}
\author{
Jay Emerson
}
\note{
None
}

\seealso{
See also \code{\link{shinySkeleton}} and \code{\link{createApp}}.
}
\examples{
# None
}
\keyword{dataset}
