% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper.R
\name{use_shinyhelper}
\alias{use_shinyhelper}
\title{Function to go in ui.R to use the package}
\usage{
use_shinyhelper(...)
}
\arguments{
\item{...}{options to pass to shinyjs::useShinyjs}
}
\description{
Function to go in ui.R to use the package
}
