% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper.R
\name{watch_helpers}
\alias{watch_helpers}
\title{Function to go in server.R to use the help messages}
\usage{
watch_helpers(input, output, help_dir = "helpfiles")
}
\arguments{
\item{input}{The input object of your shiny app}

\item{output}{The output object of your shiny app}

\item{help_dir}{A character string of the directory containing your helpfiles}
}
\description{
Function to go in server.R to use the help messages
}
\examples{
\dontrun{watch_helpers(input, output, "helpfiles")}
}
