% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-export.R
\name{get_id}
\alias{get_id}
\title{Helper function to extract the id of a shiny element as a character string}
\usage{
get_id(shiny.tag)
}
\arguments{
\item{shiny.tag}{A shiny input or output object}
}
\description{
Helper function to extract the id of a shiny element as a character string
}
