% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper.R
\name{helper}
\alias{helper}
\title{Augment a shiny.tag with a question mark helper button}
\usage{
helper(shiny.tag)
}
\arguments{
\item{shiny.tag}{A shiny input or output object}
}
\description{
Augment a shiny.tag with a question mark helper button
}
\examples{
\dontrun{plotOutput(outputId = "plot") \%>\% helper()}
}
