% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_helpers.R
\name{observe_helpers}
\alias{observe_helpers}
\title{Observe Helper Action Buttons}
\usage{
observe_helpers(session = shiny::getDefaultReactiveDomain(),
  help_dir = "helpfiles")
}
\arguments{
\item{session}{The session object in your shiny app}

\item{help_dir}{A character string of the directory containing your helpfiles}
}
\description{
Function to show a modal dialog, observing each of the help icons in the app.
}
\examples{
server <- function(input, output, session){

  # use anywhere in your server.R script
  observe_helpers()
  
  # rest of server.R
  # ...
  # ...
}

}
