% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelper-package.r
\docType{package}
\name{shinyhelper}
\alias{shinyhelper}
\alias{shinyhelper-package}
\title{shinyhelper.}
\description{
Easily add markdown helpfiles to your shiny apps, by adding a single function
around a shiny input or output. This adds a small 'question mark' help button,
which presents a markdown file in a modalDialog when clicked.
}
\details{
There is also a convenient helper function to initialise a directory of markdown
files given a vector of input and output IDs.

For further reference on how to use the package, please refer to ...
}
