% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for-shiny.R
\name{use_shinyhelper}
\alias{use_shinyhelper}
\title{Function to go in ui.R to use the package - wrapper round shinyjs::useShinyjs()}
\usage{
use_shinyhelper(...)
}
\arguments{
\item{...}{options to pass to shinyjs::useShinyjs}
}
\description{
Function to go in ui.R to use the package - wrapper round shinyjs::useShinyjs()
}
\examples{
ui <- shiny::fluidPage(

  # use at top level in your ui script
  use_shinyhelper()
  
  # rest of ui.R
  # ...
  # ...
)
}
